Use a smaller integer type for arrival-time As the arrival time is somewhere in between 0 and 86,400 seconds of a day Also skip loading copying the file content into a buffer

+++ src/tup-utils/stopTimeStore.cc
@@ -9,7 +9,7 @@
    */
 void stopTimeStore::store(std::string const& trip_id,
                           std::string const& stop_id,
-                          std::int64_t arrival_time,
+                          std::uint32_t arrival_time,
                           std::string date) {
   const auto key = std::make_tuple(trip_id, stop_id, date);
   stopTimes[key] = arrival_time;
